+++ src/components/scenes/WalletDetailsScene.tsx
@@ -174,7 +174,7 @@ function WalletDetailsComponent(props: Props) {
   const assetId = coinrankListData.coins[currencyCode]
 
   const handlePressCoinRanking = useHandler(() => {
-    navigation.navigate('coinRankingDetails', { assetId, fiatCurrencyCode })
+    navigation.navigate('coinRankingDetails', { assetId })
   })
 
   const handlePressSeeAll = useHandler(() => {